fix: rename property value source registry to .service pattern and fix import order
