+++ apps/backend/src/modules/devices/devices.module.ts
@@ -67,8 +67,8 @@ import { DevicesService } from './services/devices.service';
 import { ModuleResetService } from './services/module-reset.service';
 import { PlatformRegistryService } from './services/platform.registry.service';
 import { PropertyCommandService } from './services/property-command.service';
-import { PropertyValueSourceRegistryService } from './services/property-value-source.registry';
 import { PropertyTimeseriesService } from './services/property-timeseries.service';
+import { PropertyValueSourceRegistryService } from './services/property-value-source.registry.service';
 import { PropertyValueService } from './services/property-value.service';
 import { StatsService } from './services/stats.service';
 import { ChannelControlEntitySubscriber } from './subscribers/channel-control-entity.subscriber';
